Implemented a flexible search for analyze document
The search can now find common substrings between the search term and
the search text.  To keep the results low, I throw out results where the
common substring is too small.  There's a class variable that stores the
threshold percentage, which is currently set to 75.  That is, substrings
have to have 75% of the length of the smallest string (aka, the search
term), for the match to be considered.