Fixed some bugs in the sbh query tests
First, the search term proteomics now returns no results, because of the
type filtering.  This is still a decent test case, so I left it in but I
parameterized the test to take a term and number of search results.  I
also then added a new test using MG1655 which does have some results.