Initial implementation of analyze status messages
Now when an analyze start comes in, the analysis is performed in a
separate thread.  The connection handling thread then periodically sends
an update message to the GAS code, with a progress indicator.  Once
progress hits 100%, the results are returned.

This currently works, although I have yet to write the mechanism that
updates progress between 0 and 100.