Added test case for a bug I found
I found that the culling method in IPS wasn't quite right.  I tried to
make it more efficient by not considering things that were already
matched but this could lead to situations where an earlier match with a
short result to a long result could remove the long result and leave
later smaller overlaps to the maximal result.