Added heartbeats for sessions management
Whenever a sidebar is opened, a script is injected into the HTML.  The
script creates a timer with a 1s interval that pings the GAS code.
There's also an isActiveSession function that will scan document
properties and look at the last heartbeat from any sidebar.  If there is
a heartbeat that is less than 2s old, the function returns true,
otherwise it returns false.  Any sessions that are expired are also
removed from the property store.