Restructured json schema caching, added lab ids
There's a lot of schema data defined in json files in the data catalog.
More and more we will need to be reading in this data.

It seems like extra clutter to define a class variable for each URL, and
a separate function.  I've consolidated all of the json schema caching
into one function.