Pass pageWidth into the table creation code
The problem is, despite the documentation saying that
Table.setColumnWidth and Body.getPageWidth are in points they clearly
are not.  The page width is something like 612 in my test, and the table
width is 67.  There seems to be a difference in scale and that will have
to be figured out for the page width to be very useful.